Ancestral State Reconstruction

Perform ancestral state reconstruction. These functions are all generic and will dispatch on the class of the given likelihood functions. Currently methods exist for all generics for Mk2, and marginal ancestral state reconstructions are supported for BiSSE.

asr.marginal(lik, pars, nodes=NULL, ...)
asr.joint(lik, pars, n=1, ...)
asr.stoch(lik, pars, n=1, ...) 

make.asr.marginal(lik, ...)
make.asr.joint(lik, ...)
make.asr.stoch(lik, ...)

Arguments

lik

A likelihood function.

pars

A vector of parameters, suitable for lik.

nodes

For asr.marginal only; an optional vector of nodes to return ancestral states for (using ape's index). By default, all nodes are returned.

n

The number of samples to draw from the joint distribution, or number of stochastic reconstructions to make.

...

Additional arguments passed through to future methods

Details

These three functions compute marginal, joint, and stochastic ancestral reconstructions. The make versions return functions that can efficiently be used many times over.

Value

The return values of the functions are likely to change in the near future. Watch out!

See also

Author

Richard G. FitzJohn