Verbosely Track Function Evaluations

This is for making function evaluations more verbose; big.brother returns a function that prints its arguments and return value every iteration (or every several iterations if interval is specified)

big.brother(f, interval=1, file="")

Arguments

f

A function

interval

Interval between printed evaluations

file

(Optional) file to log to. By default, logs to the console.

Warning

Using this directly on the likelihood functions may lead to unintended consequences as it affects the class attribute of the likelihood function and stop it causing the correct MLE function being dispatched. Probably best to leave this alone.

Author

Richard G. FitzJohn