Plot Marginal Distributions from MCMC

Simple plotting assistance for plotting output from MCMC runs

profiles.plot(y, col.line, col.fill, xlim=NULL, ymax=NULL, n.br=50,
              opacity=.5, xlab="Parameter estimate",
              ylab="Probability density", legend.pos=NULL,
              with.bar=TRUE, col.bg=NA, lwd=1, lines.on.top=TRUE, ...)

Arguments

y

Data frame, columns of which will be plotted as separate profiles.

col.line

Vector of colours for the lines.

col.fill

Vector of colours for the fill of the 95% most probable region of the distribution. If ommited, this will be a semi-transparent version of col.line.

xlim

X-axis limits - calculated automatically if omitted.

ymax

Y-axis upper limit - calculated automatically if omitted.

n.br

Number of breaks along the range of the data.

opacity

Opacity of the filled region (0 is transparent, 1 is fully opaque).

xlab,ylab

Axis labels for the plot.

legend.pos

String to pass to legend to position the legend (for automatic legend building based on the names of y).

with.bar

Should a bar be included that shows the CI ranges below the plot (in addition to the shading)?

col.bg

Colour to draw behind the profiles (set to "white" for nicer transparency on non-white backgrounds)

lwd

Width of lines around the profiles

lines.on.top

Draw lines around profiles on top of all profiles?

...

Additional arguments passed through to plot.

Author

Richard G. FitzJohn

Examples

## For usage, see the example in ?make.bd