Support Functions for QuaSSE Models

Utility functions for working with QuaSSE models. These provide a minimal set of state-varying functions, suitable for use with make.quasse, and simulation assistance functions for use with tree.quasse.

This is currently poorly explained!

constant.x(x, c)
sigmoid.x(x, y0, y1, xmid, r)
stepf.x(x, y0, y1, xmid)
noroptimal.x(x, y0, y1, xmid, s2)

make.linear.x(x0, x1)

make.brownian.with.drift(drift, diffusion)

Arguments

x

Character state

c

Constant.

y0

y value at very small x (limit as x tends to negative infinity)

y1

y value at very large x (limit as x tends to infinity). For noroptimal.x, this is the y value at xmid.

xmid

Midpoint (inflection point) of sigmoid or step function

r

Rate at which exponential decay occurs or sigmoid changes - higher values are steeper

s2

Variance of the normal distribution specified by noroptimal.x.

x0

Lower x limit for the linear function: y will take value at x0 for all x smaller than this

x1

Upper x limit for the linear function: y will take value at x1 for all x greater than this

drift

Rate of drift

diffusion

Rate of diffusion (positive)

Details

The linear function returned by (make.linear.x) will go to zero wherever negative. This may not always be desired, but is required for valid likelihood calculations under QuaSSE.

Author

Richard G. FitzJohn